audio_fraud/app/requirements.txt
# python==3.7.16
keras==2.11.0
librosa==0.9.2
matplotlib==3.3.4
numpy==1.21.6
pandas==1.1.5
psycopg2==2.9.9
scikit_learn==1.0.2
tensorflow==2.11.0
python-decouple==3.8